requirements-file.txt
Flask==2.1.0
gunicorn==20.1.0
requests==2.26.0
PyJWT==2.3.0
openai==0.27.0
anthropic==0.3.0
google-generativeai==0.1.0
APScheduler==3.9.1
SQLAlchemy==1.4.31
psycopg2-binary==2.9.3
